regen with std::min and std::max-aware sql2cpp

+++ src/pluto_speech/Table_Alias.cpp
@@ -217,7 +217,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[61];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Keyword_Input.c_str(), (unsigned long) min(30,m_Keyword_Input.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Keyword_Input.c_str(), (unsigned long) min((size_t)30,m_Keyword_Input.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -231,7 +231,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[61];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_Keyword_Meaning.c_str(), (unsigned long) min(30,m_Keyword_Meaning.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_Keyword_Meaning.c_str(), (unsigned long) min((size_t)30,m_Keyword_Meaning.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -297,7 +297,7 @@ values_list_comma_separated = values_list_comma_separated + pRow->PK_Alias_asSQL
 		{
 			
 			
-			long int id	= (long int) mysql_insert_id(database->m_pMySQL);
+			long int id = (long int) mysql_insert_id(database->m_pMySQL);
