+++ src/pluto_telecom/Table_OutgoingLog.cpp
@@ -345,7 +345,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[61];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_PhoneNumber.c_str(), (unsigned long) min(30,m_PhoneNumber.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_PhoneNumber.c_str(), (unsigned long) min((size_t)30,m_PhoneNumber.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -437,7 +437,7 @@ if (is_null[10])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -516,7 +516,7 @@ values_list_comma_separated = values_list_comma_separated + pRow->PK_OutgoingLog
 		{
 			
 			
-			long int id	= (long int) mysql_insert_id(database->m_pMySQL);
+			long int id = (long int) mysql_insert_id(database->m_pMySQL);
 		
 			if (id!=0)
 pRow->m_PK_OutgoingLog=id;
