+++ src/pluto_telecom/Table_Users.cpp
@@ -258,7 +258,7 @@ if (is_null[1])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_SyncSettings.c_str(), (unsigned long) min(65535,m_SyncSettings.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_SyncSettings.c_str(), (unsigned long) min((size_t)65535,m_SyncSettings.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -324,7 +324,7 @@ if (is_null[6])
 return "NULL";
 
 char *buf = new char[29];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min(14,m_psc_mod.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_psc_mod.c_str(), (unsigned long) min((size_t)14,m_psc_mod.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -403,7 +403,7 @@ values_list_comma_separated = values_list_comma_separated + pRow->EK_Users_asSQL
 		{
 			
 			
-			long int id	= (long int) mysql_insert_id(database->m_pMySQL);
+			long int id = (long int) mysql_insert_id(database->m_pMySQL);
