+++ src/pluto_telecom/Table_psc_telecom_bathdr.cpp
@@ -208,7 +208,7 @@ if (is_null[1])
 return "NULL";
 
 char *buf = new char[33];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_IPAddress.c_str(), (unsigned long) min(16,m_IPAddress.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_IPAddress.c_str(), (unsigned long) min((size_t)16,m_IPAddress.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -222,7 +222,7 @@ if (is_null[2])
 return "NULL";
 
 char *buf = new char[39];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_date.c_str(), (unsigned long) min(19,m_date.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_date.c_str(), (unsigned long) min((size_t)19,m_date.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -236,7 +236,7 @@ if (is_null[3])
 return "NULL";
 
 char *buf = new char[131071];
-mysql_real_escape_string(table->database->m_pMySQL, buf, m_comments.c_str(), (unsigned long) min(65535,m_comments.size()));
+mysql_real_escape_string(table->database->m_pMySQL, buf, m_comments.c_str(), (unsigned long) min((size_t)65535,m_comments.size()));
 string s=string()+"\""+buf+"\"";
 delete[] buf;
 return s;
@@ -302,7 +302,7 @@ values_list_comma_separated = values_list_comma_separated + pRow->PK_psc_telecom
 		{
 			
 			
-			long int id	= (long int) mysql_insert_id(database->m_pMySQL);
+			long int id = (long int) mysql_insert_id(database->m_pMySQL);
 		
 			if (id!=0)
 pRow->m_PK_psc_telecom_bathdr=id;
